backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12
anthropic>=0.40.0
google-genai>=1.0.0
requests>=2.32.0
python-dotenv>=1.0.0
pydantic>=2.11.0
